+++ sdk/trace/span.go
@@ -244,13 +244,17 @@ func (s *span) End(options ...trace.SpanOption) {
 	}
 }
 
-// RecordError will record err as a span event for this span. An additional call to
-// SetStatus is required if the Status of the Span should be set to Error, this method
-// does not change the Span status. If this span is not being recorded or err is nil
-// than this method does nothing.
+// RecordError will record err as a span event for this span.
+// this method does not change the Span status in default.
+// If you want to change Span status, pass WithStatus as opts.
+// this metod does nothing If this span is not being recorded or err is nil.
 func (s *span) RecordError(err error, opts ...trace.EventOption) {
 	if s == nil || err == nil || !s.IsRecording() {
 		return
+	}
+	c := trace.NewEventConfig(opts...)
+	if c.WithStatus {
+		s.SetStatus(codes.Error, "")
 	}
 
 	opts = append(opts, trace.WithAttributes(